+++ Tests/GeoSpec/GeoTrackerSpec.swift
@@ -8,7 +8,7 @@ import SQLite3
 import Foundation
 import UIKit
 import CoreLocation
-@testable import RAnalytics
+@testable import RakutenAnalytics
 #if canImport(RAnalyticsTestHelpers)
 import RAnalyticsTestHelpers
 #endif