+++ Tests/RAnalyticsTestHelpers/RAnalyticsRATExpecter.swift
@@ -1,7 +1,7 @@
 import Foundation
 import Quick
 import Nimble
-@testable import RAnalytics
+@testable import RakutenAnalytics
 
 // MARK: - RAnalyticsRATExpecter