+++ Tests/Unit/RAnalyticsExternalCollectorSpec.swift
@@ -4,7 +4,7 @@
 import Quick
 import Nimble
 import UIKit
-@testable import RAnalytics
+@testable import RakutenAnalytics
 #if canImport(RAnalyticsTestHelpers)
 import RAnalyticsTestHelpers
 #endif