+++ Tests/Unit/RAnalyticsStateSpec.swift
@@ -2,7 +2,7 @@ import Quick
 import Nimble
 import CoreLocation
 import UIKit
-@testable import RAnalytics
+@testable import RakutenAnalytics
 #if canImport(RAnalyticsTestHelpers)
 import RAnalyticsTestHelpers
 #endif