+++ Tests/Unit/ReferralAppModelSpec.swift
@@ -5,7 +5,7 @@
 import Quick
 import Nimble
 import Foundation
-@testable import RAnalytics
+@testable import RakutenAnalytics
 #if canImport(RAnalyticsTestHelpers)
 import RAnalyticsTestHelpers
 #endif